backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
cffi==1.17.1
cryptography==43.0.1
Django==4.2.13
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
sqlparse==0.5.0
typing-extensions==4.12.2
